src/packager/pyproject.toml
[tool.poetry]
name = "packager"
version = "0.1.0"
description = ""
authors = ["Dr. Ernie Prabhakar <ernest@quiltdata.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
quiltcore = "^0.5.1"
gsalib = "^1.0.6"
universal-pathlib = "^0.1.4"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
mypy = "^1.7.1"
pytest-watcher = "^0.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
